Fixed up namespace structure
