+++ Source/ACE/ACE.csproj
@@ -97,14 +97,14 @@
     <Compile Include="Command\Handlers\HelpCommands.cs" />
     <Compile Include="Entity\Adjacency.cs" />
     <Compile Include="Entity\BroadcastAction.cs" />
-    <Compile Include="Entity\CollidableObject.cs" />
-    <Compile Include="Entity\Container.cs" />
-    <Compile Include="Entity\Creature.cs" />
-    <Compile Include="Entity\Door.cs" />
+    <Compile Include="Entity\Objects\CollidableObject.cs" />
+    <Compile Include="Entity\Objects\Container.cs" />
+    <Compile Include="Entity\Objects\Creature.cs" />
+    <Compile Include="Entity\Objects\Door.cs" />
     <Compile Include="Entity\Events\BroadcastEventArgs.cs" />
     <Compile Include="Entity\Events\ChatMessageArgs.cs" />
-    <Compile Include="Entity\Lifestone.cs" />
-    <Compile Include="Entity\Monster.cs" />
+    <Compile Include="Entity\Objects\Lifestone.cs" />
+    <Compile Include="Entity\Objects\Monster.cs" />
     <Compile Include="Entity\MovementData.cs" />
     <Compile Include="Entity\Events\SpellCastArgs.cs" />
     <Compile Include="Entity\Landblock.cs" />
@@ -115,10 +115,10 @@
     <Compile Include="Entity\ModelTexture.cs" />
     <Compile Include="Entity\ModelData.cs" />
     <Compile Include="Entity\PhysicsData.cs" />
-    <Compile Include="Entity\Portal.cs" />
-    <Compile Include="Entity\SpellLikeEffect.cs" />
-    <Compile Include="Entity\SummonedPortal.cs" />
-    <Compile Include="Entity\UsableObject.cs" />
+    <Compile Include="Entity\Objects\Portal.cs" />
+    <Compile Include="Entity\Objects\SpellLikeEffect.cs" />
+    <Compile Include="Entity\Objects\SummonedPortal.cs" />
+    <Compile Include="Entity\Objects\UsableObject.cs" />
     <Compile Include="Factories\CommonObjectFactory.cs" />
     <Compile Include="Factories\CorpseObjectFactory.cs" />
     <Compile Include="Factories\GenericObjectFactory.cs" />
@@ -165,8 +165,8 @@
     <Compile Include="Network\GameAction\GameActionType.cs" />
     <Compile Include="Network\Enum\PhysicsDescriptionFlag.cs" />
     <Compile Include="Network\Enum\WeenieHeaderFlags.cs" />
-    <Compile Include="Entity\Player.cs" />
-    <Compile Include="Entity\WorldObject.cs" />
+    <Compile Include="Entity\Objects\Player.cs" />
+    <Compile Include="Entity\Objects\WorldObject.cs" />
     <Compile Include="Managers\WorldManager.cs" />
     <Compile Include="Network\Enum\CharacterError.cs" />
     <Compile Include="Network\GameAction\GameAction.cs" />
@@ -202,20 +202,20 @@
     <Compile Include="Network\GameMessages\Messages\GameMessagePutObjectIn3d.cs" />
     <Compile Include="Network\GameMessages\Messages\GameMessagePutObjectInContainer.cs" />
     <Compile Include="Network\GameMessages\Messages\GameMessageUpdateInstanceId.cs" />
-    <Compile Include="Network\Handlers\PlayerTeleport.cs" />
-    <Compile Include="Network\Handlers\PlayerOptions.cs" />
-    <Compile Include="Network\Handlers\PlayerStatistics.cs" />
-    <Compile Include="Network\Handlers\PlayerGeneral.cs" />
-    <Compile Include="Network\Handlers\PlayerChat.cs" />
-    <Compile Include="Network\Handlers\PlayerCombat.cs" />
-    <Compile Include="Network\Handlers\PlayerHousing.cs" />
-    <Compile Include="Network\Handlers\PlayerPhysics.cs" />
-    <Compile Include="Network\Handlers\PlayerAllegiance.cs" />
-    <Compile Include="Network\Handlers\PlayerFriends.cs" />
-    <Compile Include="Network\Handlers\PlayerHandlers.cs" />
-    <Compile Include="Network\Handlers\PlayerAdmin.cs" />
-    <Compile Include="Network\Handlers\PlayerObjects.cs" />
-    <Compile Include="Network\Handlers\SessionHandlers.cs" />
+    <Compile Include="Entity\Objects\PlayerTeleport.cs" />
+    <Compile Include="Entity\Objects\PlayerOptions.cs" />
+    <Compile Include="Entity\Objects\PlayerStatistics.cs" />
+    <Compile Include="Entity\Objects\PlayerGeneral.cs" />
+    <Compile Include="Entity\Objects\PlayerChat.cs" />
+    <Compile Include="Entity\Objects\PlayerCombat.cs" />
+    <Compile Include="Entity\Objects\PlayerHousing.cs" />
+    <Compile Include="Entity\Objects\PlayerPhysics.cs" />
+    <Compile Include="Entity\Objects\PlayerAllegiance.cs" />
+    <Compile Include="Entity\Objects\PlayerFriends.cs" />
+    <Compile Include="Entity\Objects\PlayerHandlers.cs" />
+    <Compile Include="Entity\Objects\PlayerAdmin.cs" />
+    <Compile Include="Entity\Objects\PlayerObjects.cs" />
+    <Compile Include="Network\SessionHandlers.cs" />
     <Compile Include="Network\Motion\MotionItem.cs" />
     <Compile Include="Network\GameMessages\Messages\GameMessageAutonomousPosition.cs" />
     <Compile Include="Network\GameMessages\Messages\GameMessageCharacterCreateResponse.cs" />
@@ -250,10 +250,10 @@
     <Compile Include="Network\GameMessages\Messages\GameMessageUpdatePosition.cs" />
     <Compile Include="Network\GameMessages\Messages\GameMessageSystemChat.cs" />
     <Compile Include="Network\GameMessages\Messages\GameMessageTurbineChat.cs" />
-    <Compile Include="Network\Handlers\SessionCharacterHandler.cs" />
-    <Compile Include="Network\Handlers\SessionDDDHandler.cs" />
-    <Compile Include="Network\Handlers\SessionFriendsOldHandler.cs" />
-    <Compile Include="Network\Handlers\SessionTurbineChatHandler.cs" />
+    <Compile Include="Network\SessionCharacterHandler.cs" />
+    <Compile Include="Network\SessionDDDHandler.cs" />
+    <Compile Include="Network\SessionFriendsOldHandler.cs" />
+    <Compile Include="Network\SessionTurbineChatHandler.cs" />
     <Compile Include="Network\Extensions.cs" />
     <Compile Include="Network\GameMessageGroup.cs" />
     <Compile Include="Network\Motion\GeneralMotion.cs" />
