+++ Source/ACE/Entity/Objects/CollidableObject.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ACE.Entity
+namespace ACE.Entity.Objects
 {
     public abstract class CollidableObject : WorldObject
     {
