+++ Source/ACE/Entity/Objects/Creature.cs
@@ -15,7 +15,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace ACE.Entity
+namespace ACE.Entity.Objects
 {
     public class Creature : Container
     {
