+++ Source/ACE/Entity/Objects/PlayerCombat.cs
@@ -2,22 +2,22 @@
 using ACE.Entity;
 using ACE.Entity.Enum;
 using ACE.Network;
+using ACE.Network.Enum;
 using ACE.Network.GameAction;
-using ACE.Network.GameEvent.Events;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ACE.Entity
+namespace ACE.Entity.Objects
 {
     public partial class Player
     {
-        [GameAction(GameActionType.PingRequest)]
-        private void PingRequestAction(ClientMessage message)
+        [GameAction(GameActionType.Suicide)]
+        private void CharacterSuicide(ClientMessage message)
         {
-            Session.EnqueueSend(new GameEventPingResponse(Session));
+           Kill();
         }
     }
 }
