+++ Source/ACE/Factories/CorpseObjectFactory.cs
@@ -1,4 +1,5 @@
 ﻿using ACE.Entity;
+using ACE.Entity.Objects;
 using ACE.Entity.Enum;
 using ACE.Managers;
 using ACE.Network.Enum;
