+++ Source/ACE/Factories/GenericObjectFactory.cs
@@ -1,5 +1,6 @@
 ﻿using ACE.Entity;
 using ACE.Entity.Enum;
+using ACE.Entity.Objects;
 using System;
 using System.Collections.Generic;
 using System.Linq;
