+++ Source/ACE/Network/GameMessages/Messages/GameMessageCreateObject.cs
@@ -1,5 +1,6 @@
 ﻿using ACE.Entity;
 using ACE.Entity.Enum;
+using ACE.Entity.Objects;
 
 namespace ACE.Network.GameMessages.Messages
 {
