+++ Source/ACE/Network/GameMessages/Messages/GameMessagePutObjectInContainer.cs
@@ -1,4 +1,5 @@
 ﻿using ACE.Entity;
+using ACE.Entity.Objects;
 
 namespace ACE.Network.GameMessages.Messages
 {
