+++ Source/ACE/Network/GameMessages/Messages/GameMessageUpdateObject.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using ACE.Entity.Objects;
 
 namespace ACE.Network.GameMessages.Messages
 {
