+++ Source/ACE/Network/SessionFriendsOldHandler.cs
